Validation NeTEx : abstraction de l’affichage des résultats
Cette PR sépare le code de validation (appels d’API enRoute, construction de
l’entité MultiValidation, stockage en base, mécanismes de polling et gestion
d’erreurs) du code qui sait interpréter les résultats de validation NeTEx.

Cette séparation sera importante pour ensuite faire évoluer la représentation
des résultats tout en état capable d’afficher les anciennes validations.

Cette PR introduit notamment l’interface `ResultsAdapter`, abstraction des
fonctionnalités d’inspection des résultats (essentiellement utilisées dans les
écrans). Cette interface est le compagnon de `Validator`.

Dans cette PR il n’y a donc qu’une implémentation de `ResultsAdapter`, `V0_1_0`,
chargée d’inspecter les résultats de validation présents en production à date
de cette PR.

Cette PR ne change pas le comportement de la validation NeTEx.

L’essentiel de la différence consiste donc en 2 changements :
- une bonne partie du code de `Transports.Validators.NeTEx.Validator` est extrait
  dans `Transports.Validators.NeTEx.ResultsAdapter.V0_1_0`
- les écrans sont adaptés pour utilisé l’interface `ResultsAdapter` moyennant
  une fonction de "dispatch" (`ResultsAdapter.resolve`). Cette fonction est
  triviale puisqu’une seule version est actuellement supportée. Elle est
  également un "catch-all" pour l’avenir.
