+++ .credo.exs
@@ -76,7 +76,7 @@
         {Credo.Check.Readability.MaxLineLength, priority: :low, max_length: 120},
         {Credo.Check.Readability.ModuleAttributeNames},
         {Credo.Check.Readability.ModuleDoc},
-        {Credo.Check.Readability.ModuleNames},
+        {Credo.Check.Readability.ModuleNames, ignore: [~r/Transport.Validators.NeTEx.ResultsAdapters.V/]},
         {Credo.Check.Readability.ParenthesesOnZeroArityDefs},
         {Credo.Check.Readability.ParenthesesInCondition},
         {Credo.Check.Readability.PredicateFunctionNames},
