+++ apps/transport/lib/transport_web/templates/resource/gtfs_details.html.heex
@@ -94,7 +94,7 @@ locale = get_session(@conn, :locale) %>
                 conn: @conn,
                 data_vis: @data_vis,
                 token: nil,
-                validator: @validator
+                results_adapter: @validator
               ) %>
             </nav>
             <div class="main-pane">
