+++ apps/transport/lib/transport_web/templates/resource/netex_details.html.heex
@@ -37,7 +37,7 @@
               resource: @resource,
               severities_count: @severities_count,
               validation_summary: @validation_summary,
-              validator: @validator
+              results_adapter: @results_adapter
             ) %>
           <% end %>
           <p>
