+++ apps/transport/lib/transport_web/templates/validation/show_gtfs.html.heex
@@ -28,7 +28,7 @@
           issues: @issues,
           data_vis: @data_vis,
           token: @token,
-          validator: @validator
+          results_adapter: @validator
         ) %>
       <% end %>
