+++ apps/transport/test/transport/validators/netex/results_adapters/v0_1_0_test.exs
@@ -1,0 +1,4 @@
+defmodule Transport.Validators.NeTEx.ResultsAdapters.V0_1_0Test do
+  use ExUnit.Case, async: true
+  doctest Transport.Validators.NeTEx.ResultsAdapters.V0_1_0, import: true
+end
